Revert "Merge branch 'Avenue-Staging' into Sync-Gilbert-with-Udot"
This reverts commit bc2600345b06228064925aacee6107b5cd90a7be, reversing
changes made to 1a07b9ef53ae9279f42ffa02a64ae5c5431c2181.

+++ Atspm/Application/Analysis/Plans/Plan.cs
@@ -30,7 +30,7 @@ namespace Utah.Udot.Atspm.Analysis.Plans
         public string LocationIdentifier { get; set; }
 
         /// <inheritdoc/>
-        public short PlanNumber { get; set; }
+        public int PlanNumber { get; set; }
 
         /// <inheritdoc/>
         public virtual void AssignToPlan<T>(IEnumerable<T> range) where T : IStartEndRange
