+++ Atspm/Data/Interfaces/IPlanLayer.cs
@@ -15,6 +15,8 @@
 // limitations under the License.
 #endregion
 
+using Utah.Udot.Atspm.Data.Enums;
+
 namespace Utah.Udot.Atspm.Data.Interfaces
 {
     /// <summary>
@@ -23,10 +25,8 @@ namespace Utah.Udot.Atspm.Data.Interfaces
     public interface IPlanLayer : ILocationLayer
     {
         /// <summary>
-        /// The identifier of the signal timing plan active during the aggregation period.  
-        /// Used to associate performance metrics with specific coordination patterns
-        /// or time‑of‑day schedules.
+        /// Plan number as derrived from the event parameter on <see cref="IndianaEnumerations.CoordPatternChange"/> event
         /// </summary>
-        short PlanNumber { get; set; }
+        int PlanNumber { get; set; }
     }
 }
